+++ Source/SPConstants.h
@@ -684,7 +684,7 @@ typedef NSUInteger NSCellHitResult;
 #define SPLog(fmt, ...) NSLog((@"%s:%d: " fmt), __PRETTY_FUNCTION__, __LINE__, ##__VA_ARGS__)
 
 // See http://stackoverflow.com/questions/4415524
-#define COUNT_OF(x) ((sizeof(x)/sizeof(0[x])) / ((size_t)(!(sizeof(x) % sizeof(0[x])))))
+#define COUNT_OF(x) (NSInteger)((sizeof(x)/sizeof(0[x])) / ((size_t)(!(sizeof(x) % sizeof(0[x])))))
 
 // This definition is mostly for legibility
 #ifndef ESUCCESS
